+ Denoising supported in ExtendableSheafGCNLayer
